// lib/modules/platform/types.ts
export interface PrDebugData {
  createdInVer: string;
  updatedInVer: string;
  targetBranch: string;
  labels?: string[];
}

export interface PrBodyStruct {
  hash: string;
  rawConfigHash?: string;
  rebaseRequested?: boolean;
  debugData?: PrDebugData;
}

/**
 *
 */
export interface Pr {
  bodyStruct?: PrBodyStruct;
  sourceBranch: string;
  cannotMergeReason?: string; // for reflecting platform policies which may prevent merging
  createdAt?: string;
  closedAt?: string;
  hasAssignees?: boolean;
  labels?: string[];
  number: number;
  reviewers?: string[];
  sha?: LongCommitSha;
  sourceRepo?: string;
  state: string;
  targetBranch?: string;
  title: string;
  isDraft?: boolean;
}